requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1

# Salesforce
simple-salesforce==1.12.6

# Claude AI
anthropic==0.69.0

# Environment & Config
python-dotenv==1.0.0

# LangSmith (Observability)
langsmith==0.1.147
langchain==0.3.15

# HTTP Client (for Railway health checks)
httpx==0.28.1
